chore: move getSigners and getChannelId to utils
